Call send-email APIs from compose module
Create a draft when module is instantiated, and use that to link
the message model to the gmail backend apis. In the future this
draft will be updated more frequently than just on send.

Also modify the content provider's sendDraft method signature.
It turns out the result of calling the gmail APIs doesn't return
a full message, just the fragment used to identify the draft, so
there's not much the client can do with a message object. Instead
return a status object with a success code and an optional message.

Change-Id: Ib9a1f2eab7d23bb9fb45b11e6afd182587e524aa