Change how drafts are represented in the composer data model
Removes the separate Draft FIDL struct in favor of adding
a draft ID field to the Message struct. This change
is reflected in the models Message class as well.
When threads for the special label 'DRAFT' are fetched,
also fetch the draft IDs for the resulting threads
using a separate API call.

The thread list now chooses to launch the composer
module based on the draft ID field of selected threads'
messages.

The composer module also now stores its message state,
including the draft id, in a named link.

Change-Id: I5ffaa5b3e556f8500579c65bec5a1afa65326454