Update draft contents by fetching draft from content provider
The draft contents from the content provider are now authoritative.
This allows simplification of the state management; store state
in the initial link, not a private link. The initial link's content
is used only if there is not also a draft ID in the link.

Also fix some error handling logic in the email content provider:
if the draft update call fails, previously it tried to call the
callback with an illegal null value. Now it calls the callback
with an empty draft; the lack of a draft ID is a signal to the client
that the call failed. (There is still more to be done to improve
error handling throughout.)

Finally restore a missing call to ModuleWidget.advertise.

SO-547 #done

Change-Id: I159e0a09ad43edadd149b8d7a384912453c88a43