feat(composer) Check for mailboxes before finalizing send
Add an onPreSend callback to the ComposerModel, and wire up this
callback to the _checkForRecipients method of RecipientsList.
The idioms used here are a bit awkward because the right hook is
held by a state class which isn't visible to the code that builds
the widget hierarchy. It would be nice to find a better pattern
for this kind of thing.

SO-558 #done

Change-Id: I4abc4dd392ad0bb150f1ed5dc3ee30593cf98d85