fix(interruptions): interactive message notifications
Interruption detials:

* This change creates an interruption for any new message with the
"INBOX" label.
* A new Story will be created when the interruption is tapped.
* email/thread_list will add new messages if thier label is focused.

NOTE: In the future the ideal UX will be documented and implemented so
that if an existing story can be focused (without interrupting task
flow) it will be instead of creating new stories for each interruption.
See SO-610.

Updates the content provider's method for retrieving the Gmail API
instance so that it isn't reset everytime it is accessed. This prevents
an issue where the API instance's private values tracking history
between REST requests were being lost.

The email_api package needs to be removed and the email/content_provider
should be heavily refactored. See SO-607.

SO-575 #done

Change-Id: I6518292d49fbe2a82ed0f100bd97427896d744ca